+++ tests/test_dunamai.py
@@ -86,19 +86,19 @@ def test__version__serialize__pep440():
     )
     assert (
         Version("1", epoch=2, pre=("a", 3), post=4, dev=5, commit="abc", dirty=True).serialize(
-            with_dirty=True
+            dirty=True
         )
         == "2!1a3.post4.dev5+abc.dirty"
     )
     assert (
         Version("1", epoch=2, pre=("a", 3), post=4, dev=5, commit="abc", dirty=True).serialize(
-            with_metadata=False
+            metadata=False
         )
         == "2!1a3.post4.dev5"
     )
     assert (
         Version("1", epoch=2, pre=("a", 3), post=4, dev=5, commit="abc", dirty=True).serialize(
-            with_metadata=False, with_dirty=True
+            metadata=False, dirty=True
         )
         == "2!1a3.post4.dev5"
     )
@@ -125,19 +125,19 @@ def test__version__serialize__semver():
     assert (
         Version(
             "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True
-        ).serialize(with_dirty=True, style=style)
+        ).serialize(dirty=True, style=style)
         == "0.1.0-epoch.2.alpha.3.post.4.dev.5+abc.dirty"
     )
     assert (
         Version(
             "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True
-        ).serialize(with_metadata=False, style=style)
+        ).serialize(metadata=False, style=style)
         == "0.1.0-epoch.2.alpha.3.post.4.dev.5"
     )
     assert (
         Version(
             "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True
-        ).serialize(with_metadata=False, with_dirty=True, style=style)
+        ).serialize(metadata=False, dirty=True, style=style)
         == "0.1.0-epoch.2.alpha.3.post.4.dev.5"
     )
 
@@ -165,19 +165,19 @@ def test__version__serialize__pvp():
     assert (
         Version(
             "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True
-        ).serialize(with_dirty=True, style=style)
+        ).serialize(dirty=True, style=style)
         == "0.1.0-epoch-2-alpha-3-post-4-dev-5-abc-dirty"
     )
     assert (
         Version(
             "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True
-        ).serialize(with_metadata=False, style=style)
+        ).serialize(metadata=False, style=style)
         == "0.1.0-epoch-2-alpha-3-post-4-dev-5"
     )
     assert (
         Version(
             "0.1.0", epoch=2, pre=("alpha", 3), post=4, dev=5, commit="abc", dirty=True
-        ).serialize(with_metadata=False, with_dirty=True, style=style)
+        ).serialize(metadata=False, dirty=True, style=style)
         == "0.1.0-epoch-2-alpha-3-post-4-dev-5"
     )
 
@@ -192,59 +192,57 @@ def test__version__serialize__pvp():
     assert Version("0.1.0", pre=("rc", 3)).serialize(style=style) == "0.1.0-rc-3"
 
 
-def test__version__serialize__pep440_with_metadata():
+def test__version__serialize__pep440_metadata():
     assert Version("0.1.0").serialize() == "0.1.0"
-    assert Version("0.1.0").serialize(with_metadata=True) == "0.1.0"
-    assert Version("0.1.0").serialize(with_metadata=False) == "0.1.0"
+    assert Version("0.1.0").serialize(metadata=True) == "0.1.0"
+    assert Version("0.1.0").serialize(metadata=False) == "0.1.0"
 
     assert Version("0.1.0", pre=("a", 1), commit="abc").serialize() == "0.1.0a1"
-    assert (
-        Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=True) == "0.1.0a1+abc"
-    )
-    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=False) == "0.1.0a1"
+    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(metadata=True) == "0.1.0a1+abc"
+    assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(metadata=False) == "0.1.0a1"
 
     assert Version("0.1.0", post=1, commit="abc").serialize() == "0.1.0.post1+abc"
-    assert Version("0.1.0", post=1, commit="abc").serialize(with_metadata=True) == "0.1.0.post1+abc"
-    assert Version("0.1.0", post=1, commit="abc").serialize(with_metadata=False) == "0.1.0.post1"
+    assert Version("0.1.0", post=1, commit="abc").serialize(metadata=True) == "0.1.0.post1+abc"
+    assert Version("0.1.0", post=1, commit="abc").serialize(metadata=False) == "0.1.0.post1"
 
     assert Version("0.1.0", dev=1, commit="abc").serialize() == "0.1.0.dev1+abc"
-    assert Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=True) == "0.1.0.dev1+abc"
-    assert Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=False) == "0.1.0.dev1"
+    assert Version("0.1.0", dev=1, commit="abc").serialize(metadata=True) == "0.1.0.dev1+abc"
+    assert Version("0.1.0", dev=1, commit="abc").serialize(metadata=False) == "0.1.0.dev1"
 
 
 def test__version__serialize__semver_with_metadata():
     style = Style.SemVer
     assert Version("0.1.0").serialize(style=style) == "0.1.0"
-    assert Version("0.1.0").serialize(with_metadata=True, style=style) == "0.1.0"
-    assert Version("0.1.0").serialize(with_metadata=False, style=style) == "0.1.0"
+    assert Version("0.1.0").serialize(metadata=True, style=style) == "0.1.0"
+    assert Version("0.1.0").serialize(metadata=False, style=style) == "0.1.0"
 
     assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(style=style) == "0.1.0-a.1"
     assert (
-        Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=True, style=style)
+        Version("0.1.0", pre=("a", 1), commit="abc").serialize(metadata=True, style=style)
         == "0.1.0-a.1+abc"
     )
     assert (
-        Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=False, style=style)
+        Version("0.1.0", pre=("a", 1), commit="abc").serialize(metadata=False, style=style)
         == "0.1.0-a.1"
     )
 
     assert Version("0.1.0", post=1, commit="abc").serialize(style=style) == "0.1.0-post.1+abc"
     assert (
-        Version("0.1.0", post=1, commit="abc").serialize(with_metadata=True, style=style)
+        Version("0.1.0", post=1, commit="abc").serialize(metadata=True, style=style)
         == "0.1.0-post.1+abc"
     )
     assert (
-        Version("0.1.0", post=1, commit="abc").serialize(with_metadata=False, style=style)
+        Version("0.1.0", post=1, commit="abc").serialize(metadata=False, style=style)
         == "0.1.0-post.1"
     )
 
     assert Version("0.1.0", dev=1, commit="abc").serialize(style=style) == "0.1.0-dev.1+abc"
     assert (
-        Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=True, style=style)
+        Version("0.1.0", dev=1, commit="abc").serialize(metadata=True, style=style)
         == "0.1.0-dev.1+abc"
     )
     assert (
-        Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=False, style=style)
+        Version("0.1.0", dev=1, commit="abc").serialize(metadata=False, style=style)
         == "0.1.0-dev.1"
     )
 
@@ -252,95 +250,91 @@ def test__version__serialize__semver_with_metadata():
 def test__version__serialize__pvp_with_metadata():
     style = Style.Pvp
     assert Version("0.1.0").serialize(style=style) == "0.1.0"
-    assert Version("0.1.0").serialize(with_metadata=True, style=style) == "0.1.0"
-    assert Version("0.1.0").serialize(with_metadata=False, style=style) == "0.1.0"
+    assert Version("0.1.0").serialize(metadata=True, style=style) == "0.1.0"
+    assert Version("0.1.0").serialize(metadata=False, style=style) == "0.1.0"
 
     assert Version("0.1.0", pre=("a", 1), commit="abc").serialize(style=style) == "0.1.0-a-1"
     assert (
-        Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=True, style=style)
+        Version("0.1.0", pre=("a", 1), commit="abc").serialize(metadata=True, style=style)
         == "0.1.0-a-1-abc"
     )
     assert (
-        Version("0.1.0", pre=("a", 1), commit="abc").serialize(with_metadata=False, style=style)
+        Version("0.1.0", pre=("a", 1), commit="abc").serialize(metadata=False, style=style)
         == "0.1.0-a-1"
     )
 
     assert Version("0.1.0", post=1, commit="abc").serialize(style=style) == "0.1.0-post-1-abc"
     assert (
-        Version("0.1.0", post=1, commit="abc").serialize(with_metadata=True, style=style)
+        Version("0.1.0", post=1, commit="abc").serialize(metadata=True, style=style)
         == "0.1.0-post-1-abc"
     )
     assert (
-        Version("0.1.0", post=1, commit="abc").serialize(with_metadata=False, style=style)
+        Version("0.1.0", post=1, commit="abc").serialize(metadata=False, style=style)
         == "0.1.0-post-1"
     )
 
     assert Version("0.1.0", dev=1, commit="abc").serialize(style=style) == "0.1.0-dev-1-abc"
     assert (
-        Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=True, style=style)
+        Version("0.1.0", dev=1, commit="abc").serialize(metadata=True, style=style)
         == "0.1.0-dev-1-abc"
     )
     assert (
-        Version("0.1.0", dev=1, commit="abc").serialize(with_metadata=False, style=style)
+        Version("0.1.0", dev=1, commit="abc").serialize(metadata=False, style=style)
         == "0.1.0-dev-1"
     )
 
 
 def test__version__serialize__pep440_with_dirty():
     assert Version("0.1.0", dirty=True).serialize() == "0.1.0"
-    assert Version("0.1.0", dirty=True).serialize(with_dirty=True) == "0.1.0+dirty"
+    assert Version("0.1.0", dirty=True).serialize(dirty=True) == "0.1.0+dirty"
 
     assert Version("0.1.0", dirty=False).serialize() == "0.1.0"
-    assert Version("0.1.0", dirty=False).serialize(with_dirty=True) == "0.1.0"
+    assert Version("0.1.0", dirty=False).serialize(dirty=True) == "0.1.0"
 
-    assert Version("0.1.0", dirty=True).serialize(with_metadata=True) == "0.1.0"
-    assert (
-        Version("0.1.0", dirty=True).serialize(with_metadata=True, with_dirty=True) == "0.1.0+dirty"
-    )
+    assert Version("0.1.0", dirty=True).serialize(metadata=True) == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(metadata=True, dirty=True) == "0.1.0+dirty"
 
-    assert Version("0.1.0", dirty=True).serialize(with_metadata=False) == "0.1.0"
-    assert Version("0.1.0", dirty=True).serialize(with_metadata=False, with_dirty=True) == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(metadata=False) == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(metadata=False, dirty=True) == "0.1.0"
 
 
 def test__version__serialize__semver_with_dirty():
     style = Style.SemVer
     assert Version("0.1.0", dirty=True).serialize(style=style) == "0.1.0"
-    assert Version("0.1.0", dirty=True).serialize(with_dirty=True, style=style) == "0.1.0+dirty"
+    assert Version("0.1.0", dirty=True).serialize(dirty=True, style=style) == "0.1.0+dirty"
 
     assert Version("0.1.0", dirty=False).serialize(style=style) == "0.1.0"
-    assert Version("0.1.0", dirty=False).serialize(with_dirty=True, style=style) == "0.1.0"
+    assert Version("0.1.0", dirty=False).serialize(dirty=True, style=style) == "0.1.0"
 
-    assert Version("0.1.0", dirty=True).serialize(with_metadata=True, style=style) == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(metadata=True, style=style) == "0.1.0"
     assert (
-        Version("0.1.0", dirty=True).serialize(with_metadata=True, with_dirty=True, style=style)
+        Version("0.1.0", dirty=True).serialize(metadata=True, dirty=True, style=style)
         == "0.1.0+dirty"
     )
 
-    assert Version("0.1.0", dirty=True).serialize(with_metadata=False, style=style) == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(metadata=False, style=style) == "0.1.0"
     assert (
-        Version("0.1.0", dirty=True).serialize(with_metadata=False, with_dirty=True, style=style)
-        == "0.1.0"
+        Version("0.1.0", dirty=True).serialize(metadata=False, dirty=True, style=style) == "0.1.0"
     )
 
 
 def test__version__serialize__pvp_with_dirty():
     style = Style.Pvp
     assert Version("0.1.0", dirty=True).serialize(style=style) == "0.1.0"
-    assert Version("0.1.0", dirty=True).serialize(with_dirty=True, style=style) == "0.1.0-dirty"
+    assert Version("0.1.0", dirty=True).serialize(dirty=True, style=style) == "0.1.0-dirty"
 
     assert Version("0.1.0", dirty=False).serialize(style=style) == "0.1.0"
-    assert Version("0.1.0", dirty=False).serialize(with_dirty=True, style=style) == "0.1.0"
+    assert Version("0.1.0", dirty=False).serialize(dirty=True, style=style) == "0.1.0"
 
-    assert Version("0.1.0", dirty=True).serialize(with_metadata=True, style=style) == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(metadata=True, style=style) == "0.1.0"
     assert (
-        Version("0.1.0", dirty=True).serialize(with_metadata=True, with_dirty=True, style=style)
+        Version("0.1.0", dirty=True).serialize(metadata=True, dirty=True, style=style)
         == "0.1.0-dirty"
     )
 
-    assert Version("0.1.0", dirty=True).serialize(with_metadata=False, style=style) == "0.1.0"
+    assert Version("0.1.0", dirty=True).serialize(metadata=False, style=style) == "0.1.0"
     assert (
-        Version("0.1.0", dirty=True).serialize(with_metadata=False, with_dirty=True, style=style)
-        == "0.1.0"
+        Version("0.1.0", dirty=True).serialize(metadata=False, dirty=True, style=style) == "0.1.0"
     )